+++ src/routes/+page.svelte
@@ -1,7 +1,3 @@
-<script lang="ts" context="module">
-  export const prerender = true;
-</script>
-
 <script type="typescript">
   import App from "$lib/App.svelte";
 </script>
