+++ src/routes/about/privacy/+page.svelte
@@ -1,7 +1,3 @@
-<script lang="ts" context="module">
-  export const prerender = true;
-</script>
-
 <script lang="ts">
   import { onMount } from "svelte";
