+++ src/routes/about/privacy/+page.ts
@@ -1,0 +1,1 @@
+export const prerender = true;
